actual login happening
Now the login data gets saves, the app remembers the saved user next time it is being launched. If the tokens are not expired already, they are refreshed automatically. The logout button now works; deletes all saved user data and brings the user back to the login screen. Additionally, the fanart.tv api is required for future use (images, banners, backgrounds). For testing purposes, the function testFanart is provided.