sidebar with search, settings and logout
The pagelaout got a bit rearranges to be more flexible. All three side buttons now open a sidepanel which then shows the according content. The search button extends to an input field and the logout buton extends further to let the user confirm his action and prevent accidental signout. The search result generator is already implemented. It can be tested from the command line.