setting backend
Added global methods which can get, modify and save settings from the config file. The new config and a default preset is now added. I included a quick documentation to explain the structure.
A reset feature which replaces the config file with a preset is also added. It can be accessed from the app menu.