caching of user activity and history
The user's history of shows and episodes as well as the up next to watch episodes are now cached at startup based on the user's activity. The process is *very* unoptimized and takes in my case of testing about 16 seconds. Yep, 16 whole seconds. I know its way too long and I will work on that heavily.
The information is currently malformed in such a way that it can't be used by the dashboard to render posters from it.