card buffer
Like explained in the recently added documentation, a buffer of ±2 episodes will be requested along with the clicked on. It was a struggle to create and update the cards in the correct order but I think this is a relatively stable way of doing it.
The requesting system isn't greatly developed yet, so just episode titles and descriptions will show up. The buffer area also doesn't update yet.