log saving
The helper module now saves all log messages to a local file. This happens in both development and production mode, which has terminal logging disabled. An extension to the queue module makes the required in-time-tasking possible. The worker module is not used for anything in a while and can thus be removed.