buffer class
The buffering concept for the card overlay is now handled by a buffer class instance. It will make it easier to use for different cases where the environment can be flexibly chosen. It also makes it possible to move through the buffer and update it dynamically. This must be further developed because the indexing is not correct at this point.