buffer re-init
When opening a poster for which a buffer was already initialized before and no other poster was opened meanwhile, the buffer is being restored instead of initialized again.
The initialization of a buffer takes more time than the restoration because at the first init, the required data is read from the cache or even requested from the API. Since the buffer is saved temporarily in RAM, it is faster to access than the cache which is stored on the hard drive.