splitting modules
Moving most of the functions involved in the api data retrieving system into new modules, the new files are a lot smaller.
A problem arised where the debugLog function logs messages to the dev console instead of the terminal even when the importing doesn't happen in the renderer. Can be solved by channeling the messages over ipc into the main thread.