fixing deployment issues
Fixed a lot of issues that happened during and after building the app files. The packed app now runs as expected.
Now using the local $HOME directory for configuration files and temporary storage.
The app entry has to be cleaned up and really needs a better structure but that has to be done later.