more cache and config
The modifications do not change behavior of the app or add any new feature. Cleaned up the main app file and added documentation to the cache and config related functions. 
Removed the cache location property in the config. This was not used anymore and would not make sense to use in the future as the files were moved out of the app.