here we go again
Major problems resulted in frustration and we decided to partly restart this project. As no actual release was completed yet, we don't expect problems on other people's ends. This new branch will help to kickstart the changes we planned and will later be merged into the usual branches for further development. Fingers crossed!