removing auto-reload
When an error happened in the login screen, the page was automatically reloaded before. This didn't allow the user to get more information about what was going on. Failure handler is now changing so that it is not reloading but reporting the reason for the error. Reloading or other further actions are now tasks for the frontend/user.