fixed initial page load
Fixing an issue that prevented the loading screen to show up when starting the app with a user signed in.
In commit f7cc6318e872cb93bd81a90a64749bad03262b5a, the initial loading screen was removed but as a consequence, no other page would be opened until the main page.