moved filter to renderer
Moving the methods to filter hidden shows from the show list to the renderer API. Having a separate method in the backend for the filtered list would create a new cache file which would contain mostly duplicates. In future, there will likely be more filters which don't have to be cached individually.